+++ lib/wechat/shake_around.rb
@@ -1,5 +1,13 @@
 require 'wechat/shake_around/version'
 
+require 'wechat/shake_around/common'
+require 'wechat/shake_around/shaking'
+require 'wechat/shake_around/beacon'
+
+require 'wechat/shake_around/material'
+require 'wechat/shake_around/icon'
+require 'wechat/shake_around/license'
+
 module Wechat
   module ShakeAround
     # Your code goes here...