+++ CHANGELOG.md
@@ -1,1 +1,11 @@
 # Wechat Shake Around Change Log 微信摇周边库变更日志
+
+## v0.1
+1. Common module
+2. Beacon wrapper class
+3. Shaking wrapper class
+
+## v0.2
+1. Material module
+2. Icon wrapper class
+3. License wrapper class